Add files via upload
Changes to the code made and the connections.  To make the installation process easy, I have migrated the Sharepoint list to a local collection so that you don't have to create a new list as part of the installation process.

As usual, if there any problems - please let me know! 😊
Kristine